requirements.txt
Jinja2>=3.1.2
python-slugify>=8.0.1
python-dotenv>=1.0.0
pyairtable>=2.1.0
Markdown>=3.5
MarkupSafe>=2.1.3
